# companion/openapi.yaml
openapi: 3.1.0
info:
  title: Companion API
  version: 0.1.0
  description: |
    API del empleado IA central del ecosistema.

    Companion acepta API key o Bearer JWT interno. Cuando hay principal
    autenticado, `org_id` se interpreta como customer org/work context: la
    empresa cliente donde trabaja el empleado IA. Headers como `X-Org-ID` y
    `X-User-ID` quedan como compatibilidad de dev/test, no como fuente canonica
    productiva.
servers:
  - url: http://localhost:18085
security:
  - apiKeyAuth: []
  - bearerAuth: []
paths:
...
  /v1/watchers:
    get:
      summary: Lista watchers por customer org
      parameters:
        - in: query
          name: org_id
          description: Customer org/work context. Direct cross-org requiere `companion:cross_org`.
          schema: { type: string }
      responses:
        "200": { description: Watcher list }
    post:
      summary: Crea watcher
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateWatcherRequest" }
      responses:
        "201":
          description: Watcher created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Watcher" }
  /v1/watchers/{id}:
    get:
      summary: Obtiene watcher
      parameters:
        - $ref: "#/components/parameters/UUIDPath"
      responses:
        "200": { description: Watcher }
    patch:
      summary: Actualiza watcher
      parameters:
        - $ref: "#/components/parameters/UUIDPath"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateWatcherRequest" }
      responses:
        "200":
          description: Watcher updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Watcher" }
    delete:
      summary: Borra watcher
      parameters:
        - $ref: "#/components/parameters/UUIDPath"
      responses:
        "204": { description: Deleted }
  /v1/watchers/{id}/run:
    post:
      summary: Ejecuta watcher manualmente
      parameters:
        - $ref: "#/components/parameters/UUIDPath"
      responses:
        "200": { description: Run result }
  /v1/watchers/{id}/proposals:
    get:
      summary: Lista proposals de watcher
      parameters:
        - $ref: "#/components/parameters/UUIDPath"
      responses:
        "200": { description: Proposal list }
...
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UUIDPath:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    CreateWatcherRequest:
      type: object
      required: [name, watcher_type, config]
      properties:
        org_id:
          type: string
          description: Customer org/work context. Si falta, se deriva de la identidad autenticada.
        name: { type: string }
        watcher_type:
          type: string
          enum: [capability, stale_work_orders, unconfirmed_appointments, low_stock, inactive_customers, revenue_drop]
        config:
          oneOf:
            - $ref: "#/components/schemas/CapabilityWatcherConfig"
            - type: object
              additionalProperties: true
          description: Para `watcher_type=capability`, usar `CapabilityWatcherConfig`; los tipos Pymes existentes aceptan su config historica.
        assignee_employee_id:
          type: string
          format: uuid
          description: Virtual Employee que debe recibir el trabajo generado por este Watcher. Se proyecta dentro de config en la version actual.
        enabled: { type: boolean }
    UpdateWatcherRequest:
      type: object
      properties:
        name: { type: string }
        config:
          oneOf:
            - $ref: "#/components/schemas/CapabilityWatcherConfig"
            - type: object
              additionalProperties: true
        assignee_employee_id:
          type: string
          format: uuid
          description: Enviar string vacio para limpiar el assignee.
        enabled: { type: boolean }
    CapabilityWatcherConfig:
      type: object
      required: [product_surface, connector_kind, query_operation, action_operation, action_type]
      properties:
        product_surface: { type: string }
        connector_kind: { type: string }
        query_operation: { type: string }
        query_payload: { type: object, additionalProperties: true }
        result_items_path:
          type: string
          description: Path opcional dentro del JSON de resultado donde vive la lista de items.
        condition:
          $ref: "#/components/schemas/WatcherCondition"
        action_operation: { type: string }
        action_payload_template:
          type: object
          additionalProperties: true
        action_type: { type: string }
    WatcherCondition:
      type: object
      properties:
        path: { type: string }
        operator: { type: string, enum: [exists, eq, ne, gt, gte, lt, lte, contains] }
        value: {}
    Watcher:
      type: object
      properties:
        id: { type: string, format: uuid }
        org_id: { type: string, description: Customer org/work context propietario del watcher. }
        name: { type: string }
        watcher_type: { type: string }
        config: { type: object, additionalProperties: true }
        assignee_employee_id: { type: string, format: uuid }
        enabled: { type: boolean }
        last_run_at: { type: string, format: date-time }
        last_result: { type: object, additionalProperties: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...